Initial commit
Description:

In this commit we establish two things:

1. The repository where the source code for my website will be stored
2. The domains that were purchased to hold my website. Currently they are:
	- geosteffanov.com
	- georgistefanov.com

I decided to add georgistefanov.com since it's my name and it wasn't taken
as a domain. It sounds professional, and I was a little greedy with it.

But I would prefer to use geosteffanov.com as it passes well with my email
and also my IG. In general geosteffanov is a nice shortening of my name.